context/ref: clean up
