--- context/rag_source_test.go
@@ -1,1 +1,0 @@
-package context_test
